+++ R/allPvals.R
@@ -11,6 +11,9 @@
 #' event) and 1 (event). Must not contain NAs
 #' @param multiv Univariate analysis is performed by default, however a vector 
 #' of additional variables (corresponding to colnames in srv) can be included
+#' @param statistic the statistical test to be used to compute significance.
+#' one of "logtest" (likelihood ratio test), "waldtest" (wald statistic) or "sctest"
+#' (log-rank test)
 #' @return A vector of pvalues calculated from \code{srv} ordered by 
 #' \code{measure}
 #' @examples
@@ -18,32 +21,35 @@
 #' data(nki_subset)
 #' library(Biobase)
 #' gene_vec <- exprs(nki_subset)["NM_004448", ] #ERBB2 gene id
+#'
 #' allPvals(measure = gene_vec, 
 #'     srv = pData(nki_subset), 
 #'     time = "t.dmfs", 
-#'     event = "e.dmfs")
+#'     event = "e.dmfs",
+#'     statistic = "logtest")
 #' @export
 allPvals <- function(measure, 
                      srv, 
                      time = "Time", 
                      event = "Event", 
-                     multiv = NULL) {
+                     multiv = NULL, 
+                     statistic = "logtest") {
     # House keeping 
     if (any(is.na(srv[[time]]) | 
             is.na(srv[[event]]))) stop("NAs in survival data")
     if (any(is.na(measure))) stop("NAs in measure data")
-    
+
     # Order and arrange survival object
     srv_ordered <- srv[order(measure),]
     #srv_dt <- data.table::as.data.table(srv)
     #srv_ordered <- srv_dt[order(measure)]
     srv_obj <- survival::Surv(srv_ordered[[time]], srv_ordered[[event]])
-    
+
     # Create separation list
     separations <- lapply(1:(nrow(srv) - 1), function(x) {
         rep(c(1, 2), c(x, nrow(srv) - x))
     })
-    
+
     # Create coxph formula for uni- or multivariate analysis
     if(!is.null(multiv)){
         cox_vars <- paste(c("x", multiv), collapse = " + ")
@@ -54,13 +60,11 @@ allPvals <- function(measure,
     # Calculate P-values 
     pvals <- sapply(separations, function(x) {
                     cox_form <- stats::as.formula(paste("srv_obj ~", cox_vars))
-                    summary(survival::coxph(cox_form, srv))$logtest[["pvalue"]]
+                    summary(survival::coxph(cox_form, srv))[[statistic]][["pvalue"]]
     })
     #terminal NA makes the result play well with other variables - e.g. the 
     #number of pvalues are n-1 samples, so to align HRs against samples the 
     #additional NA makes this possible
-    #pvals_adj <- p.adjust(pvals, method = "fdr")
-    #pvals_out <- c(pvals_adj, NA)   
     pvals_out <- c(pvals, NA)
     pvals_out
 }