+++ NEWS.md
@@ -1,5 +1,7 @@
 # *survivALL* v0.9.2
 
+### 09/03/2018
+ - Users can now select their choice of statistic when calculating significance in `allPvals()`
 
 ### 01/02/2018
  - updated package to better represent associated publication